backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
channels==4.3.2
asgiref==3.11.1
requests==2.32.5
python-dateutil==2.9.0.post0
twilio==9.10.2
python-dotenv==1.2.1
